Add start game dialog
This adds a new dialog that is shown when starting a new game.
It allows the user to configure the turn limit and end time
for that game. The turn limit is implemented throughout the
game but the end time is not yet implemented on the server
side and has no effect.