Implement end game
This adds support for actually ending a game. Currently this only happens
when a game with a predifined end time reaches its end. Once the last round
has come to an end we draw the leaderboard and open up the canvas for drawing
again. The points will be shown until a new game is started.