Add rainbow color easter egg (#35)
Add rainbow color easter egg

Clicking on the draw icon on the top left switches into a rainbow
color mode. We will use this activation point for now until we
decide that the mode is good enough to keep. If so we will move it
to a more prominent location in the UI.